+++ src/beignet/_periodic_displacement.py
@@ -2,17 +2,25 @@ import torch
 from torch import Tensor
 
 
-def periodic_displacement(side: Tensor, dR: Tensor) -> Tensor:
-    """Wraps displacement vectors into a hypercube.
+def periodic_displacement(box: float | Tensor, dR: Tensor) -> Tensor:
+    r"""Wraps displacement vectors into a hypercube.
 
-    Args:
-      side: Specification of hypercube size. Either,
+    Parameters
+    ----------
+    box : float or Tensor
+        Specification of hypercube size. Either:
         (a) float if all sides have equal length.
-        (b) tensor(spatial_dim) if sides have different lengths.
-      dR: Matrix of displacements; `tensor(shape=[..., spatial_dim])`.
-    Returns:
-      Matrix of wrapped displacements; `tensor(shape=[..., spatial_dim])`.
+        (b) Tensor of shape (spatial_dim,) if sides have different lengths.
+    dR : Tensor
+        Matrix of displacements with shape (..., spatial_dim).
+
+    Returns
+    -------
+    Tensor
+        Matrix of wrapped displacements with shape (..., spatial_dim).
     """
-    return (
-        torch.remainder(dR + side * torch.tensor(0.5), side) - torch.tensor(0.5) * side
+    distances = (
+        torch.remainder(dR + box * torch.tensor(0.5, dtype=torch.float32), box)
+        - torch.tensor(0.5, dtype=torch.float32) * box
     )
+    return distances
