--- src/beignet/func/_molecular_dynamics/_partition/__periodic_displacement.py
@@ -1,25 +1,0 @@
-import torch
-from torch import Tensor
-
-
-def periodic_displacement(box: float | Tensor, dR: Tensor) -> Tensor:
-    r"""Wraps displacement vectors into a hypercube.
-    Parameters
-    ----------
-    box : float or Tensor
-        Specification of hypercube size. Either:
-        (a) float if all sides have equal length.
-        (b) Tensor of shape (spatial_dim,) if sides have different lengths.
-    dR : Tensor
-        Matrix of displacements with shape (..., spatial_dim).
-
-    Returns
-    -------
-    Tensor
-        Matrix of wrapped displacements with shape (..., spatial_dim).
-    """
-    distances = (
-        torch.remainder(dR + box * torch.tensor(0.5, dtype=torch.float32), box)
-        - torch.tensor(0.5, dtype=torch.float32) * box
-    )
-    return distances
