+++ src/beignet/func/_static_field.py
